requirements/base.txt
boto3>=1.34.0
Django>=4.2,<5.0  # Keep under 5.0 for compatibility
django-cors-headers>=4.3.0
django-filter>=23.5
django-import-export>=3.3.0
django-modeltranslation>=0.18.11
django-storages>=1.14.0
django-unique-upload>=0.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf_spectacular>=0.28.0
Pillow>=10.1.0
python-decouple>=3.8
django-smart-selects>=1.6.0
django-solo>=2.1.0
django-tinymce>=3.6.0
psycopg2-binary>=2.9.9  # https://github.com/psycopg/psycopg2
django-unfold>=0.57.0
django-silk>=5.0.4
django-lifecycle>=1.1.2
django-extensions>=3.2.3
django-redis>=5.4.0
redis>=5.0.1
celery[redis]
flower
